Reuse common test logic
